requirements.txt
email-validator>=2.3.0
flask>=3.1.2
flask-login>=0.6.3
flask-migrate>=4.1.0
flask-sqlalchemy>=3.1.1
flask-wtf>=1.2.2
flask-mail>=0.10.0
fpdf2>=2.7.0,<2.8.5
gunicorn>=21.2.0
openpyxl>=3.1.2
psycopg2-binary>=2.9.11
python-dateutil>=2.9.0.post0
python-dotenv>=1.2.1
sqlalchemy>=2.0.44
werkzeug>=3.1.3
pyotp>=2.9.0
qrcode>=7.4.2
Pillow>=10.0.0
APScheduler>=3.10.4
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

